+++ riscv/src/lib.rs
@@ -35,8 +35,10 @@
 #![no_std]
 #![allow(clippy::missing_safety_doc)]
 
+pub use paste::paste;
+
 pub mod asm;
-pub(crate) mod bits;
+pub mod bits;
 pub mod delay;
 pub mod interrupt;
 pub mod register;