+++ riscv/src/register/tests.rs
@@ -1,0 +1,3 @@
+mod read_only_csr;
+mod read_write_csr;
+mod write_only_csr;